ADD: login_required // A DECORATOR
This is a new concept for me, but simple enough to understand at first perception, it creates a
kind of wrapper around any view that its applied to, and only allows access to the genuine view
if there is a user... i.e. it checks if a user is logged in, if not it will redirect to the login
page.